pyproject.toml
[tool.poetry]
name = "workshop-streamlit-kafka"
version = "0.1.0"
description = ""
authors = ["Bruno Silva <bruno.ogs@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
faker = "^33.1.0"
streamlit = "^1.41.1"
confluent-kafka = "^2.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
